Renaming files and components
